backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pymupdf==1.24.12
nltk==3.9.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-genai>=1.0.0
